CSS change - Still error
@timmd909
Edited a few html files for clarification of syntax.
Was able to add shadow for .portraitimage across all browsers.
Issue still with the contact.html page and the margin or padding causing
the social media icons to be different. I have narrowed down the cause
to be in the common.css file, under #socialmedia ul li. The
display:inline attribute appears to be giving it some issues. What else
could I use to make it display that way?