Fix!!! Praise Jesus
Anyways. Was able to figure out the fix for the Social media icons being
all fucked up. It only took a post on Reddit for someone to explain it
to me. @timmd909 My project is complete! The only thing I have to do now
is host it on a site and get it up and live. I probably wanna fix the
form thing and use PHP at some point.